chore: launch-readiness hardening for v0.3.0 public release
- Remove 40+ internal planning docs (kairo-gap, kairo-wiki, intel, etc.) to .launch/
- Delete all dev debug artifacts (clippy_*.txt, gauntlet_*.txt, dispatch scripts)
- Comprehensive .gitignore: Rust, OS, IDE, Node, Python, secrets, .launch/
- Add CONTRIBUTING.md (dev setup, tests, Waza agent guide)
- Add CODE_OF_CONDUCT.md (Contributor Covenant v2.1)
- Add SECURITY.md (vulnerability reporting, supported versions, security architecture)
- Update CHANGELOG.md with full v0.3.0 production release entry
- Rewrite THIRD_PARTY_NOTICES.md with complete 40-crate attribution table
- Rewrite README.md to launch positioning standard (above-fold promise, quickstart)
- Update phantom-core/Cargo.toml with full publish metadata (license, repo, keywords)
- Zero clippy warnings: cargo clippy --all-targets -- -D warnings exit 0
- E2E Gauntlet: 3/3 passed
- Memory benchmark: 0.9872 composite score

+++ phantom-core/src/wgpu_effects.rs
@@ -234,7 +234,7 @@ impl WgpuEffectsEngine {
         }
 
         let start = std::time::Instant::now();
-        let total_frames = (config.duration_ms * config.fps / 1000) as u32;
+        let total_frames = config.duration_ms * config.fps / 1000;
 
         info!("[WgpuEffects] Rendering {:?} transition: {} frames @ {}fps ({}x{})",
             config.effect, total_frames, config.fps, config.width, config.height);
